requirements.txt
streamlit
tensorflow
pandas
numpy
Pillow
matplotlib
gdown
